+++ Task-10_Browser-game/src/js/components/menuComponent.js
@@ -1,6 +1,6 @@
-import { HTMLTags, ItemTypes } from './render.js'
-import { Commands } from './controls.js'
-import { IndexEnumerator } from './utils.js'
+import { HTMLTags, ItemTypes } from '../render.js'
+import { Commands } from '../controls.js'
+import { IndexEnumerator } from '../utils.js'
 
 const defaultstyleClasses = {
     table: 'width_100 align_center',
